+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_3_0/UpgradeClassNames.java
@@ -14,7 +14,6 @@
 
 package com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0;
 
-import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.LoggingTimer;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.upgrade.util.Table;
@@ -30,27 +29,9 @@ import java.util.Map;
 /**
  * @author Lino Alves
  */
-public class UpgradeClassNames extends UpgradeProcess {
+public class UpgradeClassNames extends WorkflowClassNameUpgradeProcess {
 
 	@Override
-	protected void doUpgrade() throws Exception {
-		updateClassName("KaleoInstance", "className");
-		updateClassName("KaleoInstanceToken", "className");
-		updateClassName("KaleoLog", "currentAssigneeClassName");
-		updateClassName("KaleoLog", "previousAssigneeClassName");
-		updateClassName("KaleoNotificationRecipient", "recipientClassName");
-		updateClassName("KaleoTaskAssignment", "assigneeClassName");
-		updateClassName("KaleoTaskAssignmentInstance", "assigneeClassName");
-		updateClassName("KaleoTaskInstanceToken", "className");
-
-		updateWorkflowContextEntryClassName("KaleoInstance", "kaleoInstanceId");
-		updateWorkflowContextEntryClassName("KaleoLog", "kaleoLogId");
-		updateWorkflowContextEntryClassName(
-			"KaleoTaskInstanceToken", "kaleoTaskInstanceTokenId");
-		updateWorkflowContextEntryClassName(
-			"KaleoTimerInstanceToken", "kaleoTimerInstanceTokenId");
-	}
-
 	protected void updateClassName(String tableName, String columnName) {
 		try (LoggingTimer loggingTimer = new LoggingTimer(tableName)) {
 			Table table = new Table(tableName);
@@ -65,22 +46,7 @@ public class UpgradeClassNames extends UpgradeProcess {
 		}
 	}
 
-	protected void updateWorkflowContext(
-			String tableName, String primaryKeyName, long primaryKeyValue,
-			String workflowContext)
-		throws Exception {
-
-		try (PreparedStatement ps = connection.prepareStatement(
-				"update " + tableName + " set workflowContext = ? where " +
-					primaryKeyName + " = ?")) {
-
-			ps.setString(1, workflowContext);
-			ps.setLong(2, primaryKeyValue);
-
-			ps.executeUpdate();
-		}
-	}
-
+	@Override
 	protected void updateWorkflowContextEntryClassName(
 			String tableName, String primaryKeyName)
 		throws Exception {
